radar_chart_automation/requirements.txt
altgraph==0.17.5
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
iniconfig==2.3.0
kiwisolver==1.4.9
macholib==1.16.4
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pluggy==1.6.0
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
setuptools==80.10.2
six==1.17.0
